# Make Node authorizer's index authoritative for unauthorized calls