@tosi3k 2024-09-03
The linter presubmits are not happy because of using underscores in the name of the newly introduced method, `addEdge_locked`:

```
ERROR: plugin/pkg/auth/authorizer/node/graph.go:434:17: ST1003: should not use underscores in Go names; method addEdge_locked should be addEdgeLocked (stylecheck)
ERROR: func (g *Graph) addEdge_locked(from, to, destination *namedVertex) {
ERROR:                 ^
```

The naming pattern using underscores was used here because it is consistent with other methods in that file e.g. `createVertex_locked`, `addEdgeToDestinationIndex_locked`, etc.).

To maintain consistency and please the linters, I would change the method's name to `addEdgeLocked` in this PR and open a separate one just to clean up the rest of the methods from the underscores.