@tosi3k 2024-09-03
>To maintain consistency and please the linters, I would change the method's name to `addEdgeLocked` in this PR and open a separate one just to clean up the rest of the methods from the underscores.

Done: https://github.com/kubernetes/kubernetes/pull/127102.